refactor: extract AGENT_PRESETS and api-key cache out of AgentConfigurator
