+++ src/components/HomeDashboard.tsx
@@ -1,5 +1,5 @@
 import type { AgentConfig, DocTemplate, Session } from '../types'
-import { AGENT_PRESETS } from '../AgentConfigurator'
+import { AGENT_PRESETS } from '../lib/agent-presets'
 
 interface Starter {
   id: string
